getting rid of the sprintf in the message digest preparation according to the work-around; unit cell labels are moveable now (it not really possible to go same way for the basis, since it has a fixed position in the screen plane);

+++ trunk/xlib/xmacro.cpp
@@ -1108,7 +1108,7 @@ void XLibMacros::macFixHL(TStrObjList &Cmds, const TParamList &Options, TMacroEr
   // do not print a warning...
   if( xapp.XFile().GetLattice().IsGenerated() )  return;
   TAsymmUnit &au = xapp.XFile().GetAsymmUnit();
-  TEBitArray detached((uint32_t)au.AtomCount());
+  TEBitArray detached(au.AtomCount());
   for( size_t i=0; i < au.AtomCount(); i++ )  {
     TCAtom &ca = au.GetAtom(i);
     detached.Set(i, ca.IsDetached());
